Creates new details tab for storybook
