+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieDetails/components/index.ts
@@ -13,27 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * External dependencies.
- */
-import type { Meta, StoryObj } from '@storybook/react';
 
-/**
- * Internal dependencies.
- */
-
-import { CookieDetails } from '../..';
-
-const meta = {
-  title: 'Extension/CookiesPanel/CookieDetails',
-  component: CookieDetails,
-  tags: ['autodocs'],
-} as Meta<typeof CookieDetails>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Primary: Story = {
-  args: {},
-};
+export { default as Details } from './details';
